- Update Some Code

+++ tool.py
@@ -2469,7 +2469,7 @@ def unpack(chose, form: any = None):
             lpunpack.unpack(os.path.join(work, "super.img"), work, chose)
             if os.access(work + "system_a.img", os.F_OK):
                 for wjm in os.listdir(work):
-                    if wjm.endswith('_a.img'):
+                    if wjm.endswith('_a.img') and not os.path.exists(work + wjm.replace('_a', '')):
                         os.rename(work + wjm, work + wjm.replace('_a', ''))
                     if wjm.endswith('_b.img'):
                         if not os.path.getsize(work + wjm):
